chore(parser): remove unused stack type
