--- pkg/parser/parser_storedict.go
@@ -1,27 +1,0 @@
-package parser
-
-import (
-	"github.com/bytesparadise/libasciidoc/pkg/configuration"
-)
-
-// extra methods on the generated parser's `storeDict` type
-
-const attributesKey = "attributes"
-
-const usermacrosKey = "user_macros"
-
-func (c storeDict) getAttributes() interface{} {
-	if s, ok := c[attributesKey].(*stack); ok {
-		return s.get()
-	}
-	return nil
-}
-
-func (c storeDict) hasUserMacro(name string) bool {
-	if macros, exists := c[usermacrosKey].(map[string]configuration.MacroTemplate); exists {
-		_, exists := macros[name]
-		return exists
-	}
-	// log.Debugf("no user macro registered")
-	return false
-}
